cleanup testing code

+++ config/services.php
@@ -53,7 +53,7 @@ return [
         'pagination_per_page' => env('WEBHOOKS_PAGINATION_PER_PAGE', 20),
         'allowed_item_types' => env('WEBHOOK_ALLOWED_ITEM_TYPES', 'Movie,Episode,Audio'),
         'max_entries' => env('WEBHOOK_MAX_ENTRIES', 100),
-        'image_cache_ttl' => env('IMAGE_CACHE_TTL_MINUTES', 1440),
+        'image_cache_ttl' => (int) env('IMAGE_CACHE_TTL_MINUTES', 1440),
     ],
 
     'emby' => [
